backend/requirements.txt
Flask
Flask-SocketIO
Flask-Cors
pyserial